feat(runtime): represent every Pine int as a float
A Pine int is a static type only: at runtime every int-typed value is a
double. The int-ness lives in the type inference and the overload pin, the
values are native floats, and the numeric na of an int is the same nan a
float has. Truncation happens at the consumer, storage preserves the value.

Conversion boundaries:
- Series[int] / Persistent[int] slots store floats (SeriesImpl as_float);
  a Persistent[int] literal initializer in a script module becomes a float
- bar_index, last_bar_index, last_bar_time and the syminfo int fields are
  published as floats; input.int and the generic input() hand back floats
- int-returning lib functions (time, year, dayofweek, array.size,
  str.length, ta.barssince, strategy trade counters, color.r/g/b, ...)
  convert at the return boundary
- int() in a script is the Pine cast (safe_int, a float); in a @pyne lib
  module it is the lib's own truncation (native_int, a Python int); every
  non-literal range() argument is truncated with native_int
- pine_range takes a value-based fast path yielding floats, never truncates
- NA(int) and na_int are the native nan; NA.__reduce__ keeps pickling exact
- math.round_to_mintick multiplies its exact tie-break by the native int
  tick pair, so the float pricescale/minmove cannot flip a measured tie
- inline_series leaves the offset to the series (an na offset is bar 0),
  str.repeat and color.r/g/b answer a numeric na by value
- the CSV writer prints integral floats without a fraction (the way the
  TradingView export writes bar_index and Volume) and keeps the sign of
  -0.0; the reader parses extra fields as floats

ta.median always returns the mean of the two middles, ta.change takes the
bool branch first, ta highest/lowest address their rings with the
truncated bar index. lib_types.json and module_properties.json are
regenerated; timeframe._in_seconds is the native-int core face of
in_seconds.

Tests: runtime float-only coverage, script/lib lowering and range
rewriting, NA pickle round trip, CSV integral formatting; goldens and
expectations updated for the float representation.

+++ src/pynecore/types/na.pyi
@@ -83,6 +83,6 @@ class NA(Generic[T]):
 # interned constants (instead of an NA(float) call) would false-positive.
 # (na_float genuinely IS a float now — the interned native nan.)
 na_float: float
-na_int: int
+na_int: float
 na_str: str
 na_bool: bool
